delete file t_UserController

+++ src/main/java/com/example/skyfast_2_0/entity/Airplane.java
@@ -38,15 +38,15 @@ public class Airplane {
 
     @NotNull
     @Column(name = "total_length", nullable = false)
-    private Integer totalLength;
+    private Float totalLength;
 
     @NotNull
     @Column(name = "wingspan", nullable = false)
-    private Integer wingspan;
+    private Float  wingspan;
 
     @NotNull
     @Column(name = "height", nullable = false)
-    private Integer height;
+    private Float  height;
 
     @Size(max = 255)
     @NotNull
